+++ util/containerutil/podman.go
@@ -27,8 +27,11 @@ func NewPodmanShellFrontend(ctx context.Context, cfg *FrontendConfig) (Container
 			binaryName:              "podman",
 			runCompatibilityArgs:    []string{"--security-opt", "unmask=/sys/fs/cgroup"},
 			globalCompatibilityArgs: make([]string, 0),
+			Console:                 cfg.Console,
 		},
 	}
+	// TODO: Find a cleaner way to pass down information to the shellFrontend
+	fe.FrontendInformation = fe.Information
 
 	output, err := fe.commandContextOutput(ctx, "info", "--format={{.Host.Security.Rootless}}")
 	if err != nil {