+++ CHANGELOG.md
@@ -14,6 +14,10 @@ All notable changes to [Earthly](https://github.com/earthly/earthly) will be doc
   fallback to `/usr/share/zsh/site-functions`.
 - `earthly sat select` with no arguments now prints the current satellite and the usage text.
 
+### Fixed
+
+- Fixed Earthly failing when using a remote docker host from a machine with an incompatible architecture. [#1895](https://github.com/earthly/earthly/issues/1895)
+
 ## v0.6.23 - 2022-09-06
 
 ### Fixed